+++ index.js
@@ -1,0 +1,2 @@
+export {EtoolsDropdown} from './EtoolsDropdown.js';
+export {EtoolsDropdownMulti} from './EtoolsDropdownMulti.js';
